requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
Flask-Limiter==3.5.0
flask-restx==1.2.0
python-socketio==5.9.0
redis==5.0.1
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.1
xgboost==1.7.4
tensorflow==2.12.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
plotly==5.14.0
pytest==7.3.1
gunicorn==20.1.0
schedule==1.2.0
optuna==3.5.0
eventlet==0.33.3
tabulate==0.9.0
lightgbm==4.1.0
catboost==1.2.3